static/old_code/dev_requirements.txt
pyyaml == 6.0.1
pygame == 2.5.1